requirements.txt
pandas>=2.0
pybaseball>=2.3
supabase>=2.0
python-dotenv>=1.0
requests>=2.31
